Remove ADDITIONAL_RULES option
If we do need this, we should implement it via a more generalised config e.g. a
`.spm-uiuh-config` file, where we can add additional options, rather than adding
lots of command-line args.

This way also helps keep the batch files and shell scripts as simple, clean and
static as possible, which is important because those files are manual downloads
for the customer.